requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
Pillow>=8.3.0
opencv-python>=4.5.0
albumentations>=1.1.0
tqdm>=4.62.0
tensorboard>=2.7.0
PyYAML>=6.0
wandb>=0.12.0
pandas>=1.3.0
h5py>=3.6.0
timm>=0.5.0 